refactor(core): consolidate packages into core module
streamline package structure by moving classes into the "core" module for better modularity and maintainability. updated imports and references accordingly.

BREAKING CHANGE: reorganized package structure; update imports where applicable.
